requirements.txt
functions-framework==3.*
Flask==3.0.3
requests==2.31.0
flask-cors==4.0.0
psycopg2-binary==2.9.9